requirements.txt
python-telegram-bot==13.15
supabase==1.2.0
APScheduler==3.6.3
python-dotenv==1.0.1